change when criteria are created during studypopulation API requests
previously criteria were created/looked up before the rest of the studypopulation was.
Moving this to after the studypop creation/update makes much of the code more
straightforward (and if criteria throw an error the whole transaction will still roll back,
so there is not much effective difference in that case).