alter metric_unit handling in exposure API to disallow new metric creation
Previously had some custom code to allow callers to specify existing metric_unit
by id, by name, and to create a metric_unit on the fly if an existing one with that
name could not be found. Per code review comments, disabling this; also greatly simplifies
the code as now it can just use FlexibleDBLinkedChoiceField.